done api

+++ database/migrations/2025_05_09_021624_create_class_students_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('class_students', function (Blueprint $table) {
-        $table->unsignedInteger('class_id');
+        $table->unsignedBigInteger('class_id');
         $table->unsignedBigInteger('student_id');
         $table->timestamp('created_at')->useCurrent();
         $table->primary(['class_id', 'student_id']);
